// src/lib/api/generated/models/updateCartPayload.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Marketplace API
 * API for Village Website & Marketplace
 * OpenAPI spec version: 1.0.0
 */
import type { WeightOz } from './weightOz';

export interface UpdateCartPayload {
  quantityOz?: WeightOz;
  /** Whether this item should be updated as a recurring subscription */
  isSubscription?: boolean;
}